+++ lib/mango/core_ext/string.rb
@@ -6,11 +6,14 @@ class String
   # Convert a file name to a Sinatra-compliant template name
   #
   # @example
-  #   "blog.haml".templatize  #=> :blog
+  #   "blog.haml".templatize              #=> :blog
+  #   "blog/home.erb".templatize          #=> :"blog/home"
+  #   "page.html.liquid".templatize       #=> :"page.html"
+  #   "article/post.html.haml".templatize #=> :"article/post.html"
   #
   # @return [Symbol] A Sinatra-compliant template name
   #
   def templatize
-    File.basename(self, '.*').to_sym
+    self.rpartition(".").shift.to_sym
   end
 end