+++ spec/lib/application/routing_content_pages_spec.rb
@@ -256,6 +256,43 @@ describe Mango::Application do
 
   #################################################################################################
 
+  describe "GET /articles/index" do
+    before(:all) do
+      get "/articles/index"
+    end
+
+    it "returns 200 status code" do
+      last_response.should be_ok
+    end
+
+    it "sends the correct Content-Type header" do
+      last_response["Content-Type"].should == "text/html;charset=utf-8"
+    end
+
+    it "sends the correct body content" do
+      last_response.body.should == <<-EXPECTED
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset='utf-8' />
+    <title>layout.erb</title>
+  </head>
+  <body>
+<h1>Welcome to Mango!</h1>
+
+<p id="template">articles/page.erb</p>
+
+<div id="content">
+  <h3>/articles/index.md</h3>
+</div>
+  </body>
+</html>
+      EXPECTED
+    end
+  end
+
+  #################################################################################################
+
   describe "GET /turner%2Bhooch" do
     before(:all) do
       get "/turner%2Bhooch"