+++ spec/lib/runner_spec.rb
@@ -81,7 +81,7 @@ run Mango::Application
       File.read(expected).should == <<-EOS
 # encoding: UTF-8
 source "http://rubygems.org"
-gem "mango", "~> 0.6.1"
+gem "mango", "~> 0.6.2"
       EOS
     end