+++ CHANGES.md
@@ -12,6 +12,7 @@
 ## Dependencies
 
   * Update rack to ~> 1.2.3 [GH#63](https://github.com/ryansobol/mango/issues/63)
+  * Add therubyracer-heroku = 0.8.1.pre3 [GH#59](https://github.com/ryansobol/mango/issues/59)
 
 ## Documentation